add LiveData、move databasebuilder into WordDatabase Class

+++ app/src/main/java/com/example/myapplication/MainActivity.kt
@@ -4,6 +4,8 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.TextView
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
 import androidx.room.Room
 
 
@@ -16,6 +18,7 @@ class MainActivity : AppCompatActivity() {
     private lateinit var wordDao: WordDao
     private lateinit var wordDatabase: WordDatabase
     private lateinit var textView: TextView
+    private lateinit var allWordsLive:LiveData<List<Word>>
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -27,43 +30,48 @@ class MainActivity : AppCompatActivity() {
         updateButton = findViewById(R.id.updateButton)
         wordDatabase = Room.databaseBuilder(this, WordDatabase::class.java, "Eeko").allowMainThreadQueries().build()
         wordDao = wordDatabase.getWordDao()
+        allWordsLive=wordDao.getAllWordsLive()
         textView = findViewById(R.id.textView)
+        allWordsLive.observe(this, Observer {
+            val list: List<Word> =it
+            var text: String = ""
+            for (i in list.indices) {
+                val tempWord = list[i]
+                text = text + tempWord.id + ":" + tempWord.english + "=" + tempWord.chineseMeaning + "\n"
+            }
+            textView.text = text
+        })
 
-        updateView()
         insertButton.setOnClickListener() {
             val word1 = Word("how", "如何~")
             val word2 = Word("hello", "你好")
             val word3 = Word("taiwan", "台灣")
             wordDao.insertWords(word1,word2,word3)
-            updateView()
         }
         deleteButton.setOnClickListener(){
-            val lastWord=wordDao.getAllWords()[0]
-            wordDao.deleteWords(lastWord)
-            updateView()
+            val lastWord= allWordsLive.value?.get(0)
+            if (lastWord != null) {
+                wordDao.deleteWords(lastWord)
+            }
         }
         cleanButton.setOnClickListener(){
             wordDao.deleteAllWords()
-            updateView()
         }
         updateButton.setOnClickListener(){
-            val lastWord=wordDao.getAllWords()[0]
-            lastWord.english="404"
-            lastWord.chineseMeaning="404"
-            wordDao.updateWords(lastWord)
-            updateView()
+            val lastWord=  allWordsLive.value?.get(0)
+            if (lastWord != null) {
+                lastWord.english="404"
+            }
+            if (lastWord != null) {
+                lastWord.chineseMeaning="404"
+            }
+            if (lastWord != null) {
+                wordDao.updateWords(lastWord)
+            }
         }
 
     }
 
-    private fun updateView() {
-        val list: List<Word> = wordDao.getAllWords()
-        var text: String = ""
-        for (i in list.indices) {
-            val tempWord = list[i]
-            text = text + tempWord.id + ":" + tempWord.english + "=" + tempWord.chineseMeaning + "\n"
-        }
-        textView.text = text
-    }
+
 
 }
